requirements.txt
asgiref==3.3.2
Django==3.0.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
